Implemented LoanConstants Interface to Loan class

+++ LoanConstants.cs
@@ -8,9 +8,9 @@ namespace Loan_Application_System
 {
     public interface LoanConstants
     {
-        double shortTerm { get; }
-        double mediumTerm { get; }
-        double longTerm { get; }
+        int shortTerm { get; }
+        int mediumTerm { get; }
+        int longTerm { get; }
         string companyName { get; }
         double maxLoanAmount { get; }
     }
